Tests/packetbeat_tests/requirements.txt
allure-behave==2.8.5
allure-python-commons==2.8.5
astroid==2.2.5
attrs==19.1.0
autopep8==1.4.4
azure-devops==5.1.0b3
behave==1.2.6
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
elasticsearch==7.0.4
elasticsearch-dsl==7.0.0
func-timeout==4.3.5
idna==2.8
importlib-metadata==0.21
isodate==0.6.0
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
more-itertools==7.2.0
msrest==0.6.10
oauthlib==3.1.0
parse==1.12.1
parse-type==0.5.2
pluggy==0.13.0
pycodestyle==2.5.0
pylint==2.3.1
python-dateutil==2.8.0
python-logstash==0.4.6
requests==2.22.0
requests-oauthlib==1.2.0
rope==0.14.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
zipp==0.6.0